# python-data/XptoTracking-DataMS/swagger_server/models/get_telemetry_data_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
import re  # noqa: F401,E501
from swagger_server import util


class GetTelemetryDataRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data_id: str=None, vehicle_id: str=None, data_hora: str=None):  # noqa: E501
        """GetTelemetryDataRequest - a model defined in Swagger

        :param data_id: The data_id of this GetTelemetryDataRequest.  # noqa: E501
        :type data_id: str
        :param vehicle_id: The vehicle_id of this GetTelemetryDataRequest.  # noqa: E501
        :type vehicle_id: str
        :param data_hora: The data_hora of this GetTelemetryDataRequest.  # noqa: E501
        :type data_hora: str
        """
        self.swagger_types = {
            'data_id': str,
            'vehicle_id': str,
            'data_hora': str
        }

        self.attribute_map = {
            'data_id': 'dataId',
            'vehicle_id': 'vehicleId',
            'data_hora': 'dataHora'
        }
        self._data_id = data_id
        self._vehicle_id = vehicle_id
        self._data_hora = data_hora

    @classmethod
    def from_dict(cls, dikt) -> 'GetTelemetryDataRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GetTelemetryDataRequest of this GetTelemetryDataRequest.  # noqa: E501
        :rtype: GetTelemetryDataRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data_id(self) -> str:
        """Gets the data_id of this GetTelemetryDataRequest.


        :return: The data_id of this GetTelemetryDataRequest.
        :rtype: str
        """
        return self._data_id

    @data_id.setter
    def data_id(self, data_id: str):
        """Sets the data_id of this GetTelemetryDataRequest.


        :param data_id: The data_id of this GetTelemetryDataRequest.
        :type data_id: str
        """

        self._data_id = data_id

    @property
    def vehicle_id(self) -> str:
        """Gets the vehicle_id of this GetTelemetryDataRequest.


        :return: The vehicle_id of this GetTelemetryDataRequest.
        :rtype: str
        """
        return self._vehicle_id

    @vehicle_id.setter
    def vehicle_id(self, vehicle_id: str):
        """Sets the vehicle_id of this GetTelemetryDataRequest.


        :param vehicle_id: The vehicle_id of this GetTelemetryDataRequest.
        :type vehicle_id: str
        """

        self._vehicle_id = vehicle_id

    @property
    def data_hora(self) -> str:
        """Gets the data_hora of this GetTelemetryDataRequest.

        Date time for telemetry data  # noqa: E501

        :return: The data_hora of this GetTelemetryDataRequest.
        :rtype: str
        """
        return self._data_hora

    @data_hora.setter
    def data_hora(self, data_hora: str):
        """Sets the data_hora of this GetTelemetryDataRequest.

        Date time for telemetry data  # noqa: E501

        :param data_hora: The data_hora of this GetTelemetryDataRequest.
        :type data_hora: str
        """

        self._data_hora = data_hora